fix: исправить RequestPrivateService

+++ ewm-general-service/src/main/java/ru/practikum/ewm/general/service/adminAPI/CategoryAdminServiceImpl.java
@@ -3,6 +3,7 @@ package ru.practikum.ewm.general.service.adminAPI;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practikum.ewm.general.exception.NotAvailableException;
 import ru.practikum.ewm.general.model.Category;
 import ru.practikum.ewm.general.model.dto.CategoryDto;
 import ru.practikum.ewm.general.model.dto.NewCategoryDto;
@@ -15,6 +16,7 @@ import ru.practikum.ewm.general.repository.CategoryRepository;
 public class CategoryAdminServiceImpl implements CategoryAdminService {
 
     private final CategoryRepository categoryRepository;
+    private final EventAdminService eventAdminService;
     @Override
     public CategoryDto createCategory(NewCategoryDto dto) {
         CategoryDto categoryDto = CategoryMapper.toDto(categoryRepository
@@ -35,7 +37,12 @@ public class CategoryAdminServiceImpl implements CategoryAdminService {
     @Override
     public void deleteCategory(long id) {
         Category category = categoryRepository.get(id);
-        categoryRepository.delete(category);
-        log.info("Категория с id {} была удалена", id);
+
+        if(eventAdminService.getAllByCategory(category.getId()).isEmpty()) {
+            categoryRepository.delete(category);
+            log.info("Категория с id {} была удалена", id);
+        } else {
+            throw new NotAvailableException("event exists");
+        }
     }
 }
